Clean up the million utilities files
